// intergalactic-logistics-frontend/src/app/models/starship.model.ts
export interface StarshipInfo {
  name: string;
  model: string;
  manufacturer: string;
  costInCredits: string;
  length: string;
  maxAtmospheringSpeed: string;
  crew: string;
  passengers: string;
  cargoCapacity: string;
  hyperdriveRating: string;
  mglt: string;
  starshipClass: string;
  swapiId: string; // Numeric ID from SWAPI for lookups
}